progress

+++ src/LionFire.Trading.Abstractions/Accounts/IPosition.cs
@@ -8,19 +8,19 @@ namespace LionFire.Trading
     public interface IPosition
     {
         string Comment { get; }
-        double Commissions { get; }
-        double EntryPrice { get; }
+        decimal Commissions { get; }
+        decimal EntryPrice { get; }
         DateTime EntryTime { get; }
-        double GrossProfit { get; }
+        decimal GrossProfit { get; }
         int Id { get; }
         string Label { get; }
-        double NetProfit { get; }
-        double Pips { get; }
-        double Quantity { get; }
-        double? StopLoss { get; }
-        double Swap { get; }
+        decimal NetProfit { get; }
+        decimal Pips { get; }
+        decimal Quantity { get; }
+        decimal? StopLoss { get; }
+        decimal Swap { get; }
         string SymbolCode { get; }
-        double? TakeProfit { get; }
+        decimal? TakeProfit { get; }
         TradeKind TradeType { get; }
         long Volume { get; }
     }
